+++ NEWS.md
@@ -1,3 +1,5 @@
 # asciiSetupReader
 
 + Fixed bug where function fails if no value-labels are detected
+    + Applied both when setup file has no value-labels and if only select columns that have no value-labels
++ Added a huge number of tests